refactor/be: 회원가입 엔드포인트를 user_controller에서 auth_controller로 이동
