+++ ingq/app/macmorning.py
@@ -6,7 +6,6 @@ from core.auth_middleware import AuthMiddleware
 from core.openapi import custom_openapi
 
 from auth.interface.controller.auth_controller import router as auth_router
-from user.interface.controller.user_controller import router as user_router
 
 from dependencies.containers import Container
 from db.redis_cache import redis_cache
@@ -23,7 +22,6 @@ def create_app() -> FastAPI:
     container.init_resources()
     container.wire(
         modules=[
-            "user.interface.controller.user_controller",
             "auth.interface.controller.auth_controller",
         ]
     )
@@ -46,7 +44,6 @@ def create_app() -> FastAPI:
     CorsConfig(app=app)
 
     app.include_router(auth_router)
-    app.include_router(user_router)
 
     app.openapi = lambda: custom_openapi(app, exempt_paths)
